updated to allow single trees to be run
Made changes to the cleaning code at the start of the function so phylo objects did not throw an error and also to nest a phylo object within a multiphylo object. This is needed as trees are read into the MCMCglmm using mulTree.data$phy[[i]] which will only select a tree in a multi phylo object. Let me know if this solution is ok as its a bit of a patch job really.