I have resorted to a shitty loop for the row drop issue
The same species in my analysis keeps staying in the flipping database despite not matching the tree. I don't know why as other species seem to be dropped fine. Iv added a loop to make sure that it essentially combs over it a few times so I know its not a multiple entry thing.